last minute tweaks to improve game

+++ Project/Game.cs
@@ -9,27 +9,26 @@ namespace TheCenter.Project
 {
     public class Game : IGame
     {
-
-        Stopwatch stopwatch = new Stopwatch();
-
+        Warning countdown = new Warning(4);
         public Room CurrentRoom { get; set; }
 
         public Player CurrentPlayer { get; set; }
 
         Process myproc = new Process();
 
+        public bool running = false;
+
         public void Setup()
         {
-
             Room EntranceTunnel = new Room("EntranceTunnel", "The door behind you slams shut as you hear the tumblers lock securely.\r\nYou don’t recall what just happened aside from bright lights and a \r\nloud humming in your ears.\r\n\r\nMatter of fact, you can’t even remember what you’re doing here.\r\n\r\nYou sense greater humidity in the air. You’re in a long hallway.\r\nThe hallway extends (N)orth for nearly 100’.\r\nYou see a heavy door at the end of the hallway.\r\n\r\nThere is nothing else of interest here.", false, new Item("", ""), false);
 
-            Room AirlockTreatment = new Room("AirlockTreatment Room", "You enter this small room through a heavy door that looks to be built \r\nto maintain an air-tight seal. On the north side of the room is \r\nanother similar door. There are benches along the walls here. \r\n\r\nThere are small canvas bags on the wall near you that appear to be \r\ncarrying individual (mask)s of some sort. The door automatically closes \r\nbehind you and locks as before. A painful reminder of being trapped \r\nfills your mind. \r\n\r\nAn alarm pierces the veil of silence and spins you around as a \r\nflashing red light pierces your mind with painful intent. \r\n\r\nYou hear over the noise a countdown sequence. . .  \r\n60 seconds   59    58     57", true, new Item("mask", "mask description here"), true);
+            Room AirlockTreatment = new Room("AirlockTreatment Room", "You enter this small room through a heavy door that looks to be built \r\nto maintain an air-tight seal. An audible warning starts beeping.\r\nOn the north side of the room is \r\nanother similar door. There are benches along the walls here. \r\n\r\nThere are small canvas bags on the wall near you that appear to be \r\ncarrying individual (mask)s of some sort. The door automatically closes \r\nbehind you and locks as before. A painful reminder of being trapped \r\nfills your mind. \r\n\r\nAn alarm pierces the veil of silence and spins you around as a \r\nflashing red light pierces your mind with painful intent. \r\n\r\nYou hear over the noise a countdown sequence. . .  \r\n60 seconds   59    58     57", true, new Item("", ""), true);
 
             Room FacilitiesRoom = new Room("Facilities Room", "You enter through the last airlock. The heavy door again shuts and locks \r\nbehind you with an evil intent. \r\n\r\nYou’re in a very large room. \r\n\r\nImagine a huge room with lockers and individual showers available. \r\nMust be some sort of changing room after the flea fog you just went \r\nthrough. . . or whatever it was. Fumigation? But against what? \r\n\r\nYou notice jumpsuits of different colors hanging in various racks with \r\ngeneric ID badges of some sort on them. Each one has a TLD. . . \r\n\r\nA Thermoluminescent Dosimeter. Great. Radiation monitoring . . . but what is the source of radiation though? And what do the colors mean? \r\n\r\nYou see (bluejumpsuit)s, (blackjumpsuit)s, (redjumpsuit)s and army (greenjumpsuit)s. \r\nThere are also a few sets of (whitejumpsuit)s, each have a white (wristband) \r\npinned to them in addition to the TLD and ID. \r\n\r\nYou see a hallway to the (N)orth.", false, new Item("", ""), false);
 
             Room CentralRoom = new Room("Central Room", "You enter what appears to be the world’s largest living room. \r\nLots of seating. \r\n\r\nPeople dressed in various jumpsuits are seated leisurely all around \r\nyou in couches, individual seats, chairs and small booths that \r\nappear to be setup for mini-conferencing. These booths have a type of \r\nshade or partition that can be used to obscure the meeting. Not bad. \r\nYou decide to sit in a comfortable, high-walled but seemingly normal \r\nbooth. Something catches your attention . . .  was it a sound, \r\na (flashlight), a (laptop), or a motion at the fancy booths to the north. . .\r\n\r\nWhile you take a moment to focus on the distraction you notice activity \r\nin one of the fancy booths on the north wall. One of the two people in the booth draw the shade and you see a shift \r\nin the lighting in that (booth). Odd. \r\n\r\nYou pay no attention per se until not a minute later you see the shade \r\nin that very booth being retracted automatically, only to reveal . . .  \r\n\r\nan empty booth! \r\n\r\nA quick mental check - YEP! there were people in that booth just a few \r\nmoments ago until . . .\r\n\r\nthey drew the partition down! \r\n\r\nOK. Officially weird. \r\n\r\nAs you take a seat, you watch those booths and sure enough, people that \r\ngo into the booths and draw the partition down disappear. \r\n\r\nYou also notice people do not use the booths otherwise. \r\n\r\nYou see a hallway to the (E)ast and of course the door you originally \r\ncame through from the (S)outh.", true, new Item("", ""), false);
 
-            Room MedLabDoor = new Room("Central Room East Hallway", "You are at a frosted glass door, facing (E)ast, with the words “MedLab” \r\nin black etched into the glass. The door is locked but there is no \r\ndoorknob or keyhole. \r\n\r\nYou notice a very small pad on the wall on the side where you would \r\nexpect a doorknob.  Just then, a sliding wall extends behind you, \r\nblocking your ability to go back! \r\n\r\nYou're Trapped!\r\n", true, new Item("",""), true);  
+            Room MedLabDoor = new Room("Central Room East Hallway", "You are at a frosted glass door, facing (E)ast, with the words “MedLab” \r\nin black etched into the glass. The door is locked but there is no \r\ndoorknob or keyhole. \r\n\r\nYou notice a very small pad on the wall on the side where you would \r\nexpect a doorknob.  Just then, a sliding wall extends behind you, \r\nblocking your ability to go back! \r\n\r\nYou're Trapped!\r\n", true, new Item("", ""), true);
 
             Room MedLab = new Room("MedLab", "You enter a well lit room obviously meant for medical needs. \r\nThere are cabinets and drawers, display cases and exam areas \r\nfor up to 4 patients. \r\n\r\nOne cabinet is marked with “Defibrillator” and you see other \r\npieces of equipment such as blood pressure bands, oxygen \r\nsystems with masks, (steth)oscopes, (clip)boards, resuscitation \r\nbag and mask, a small autoclave, a computer on a desk \r\nwith typical desk items around including an odd looking small \r\nplastic (card) next to the computer, a small water system \r\nwith small drink cups, refrigerator for medical use and the \r\nobligatory (glov)es, tongue (dep)ressors, sharps container etc. \r\n\r\nThere are no other exits other than the one you walked through \r\nfrom the (W)est.", false, new Item("", ""), false);
 
@@ -72,7 +71,7 @@ namespace TheCenter.Project
             MedLab.Exits.Add("w", CentralRoom);
             myproc.StartInfo.FileName = @".\Project\Media\cmdmp3.exe";
             myproc.StartInfo.Arguments = @".\Project\Media\AWA.mp3";
-            myproc.Start();
+            // myproc.Start();
             CurrentRoom = EntranceTunnel;
             CurrentPlayer = new Project.Player();
         }
@@ -99,8 +98,9 @@ namespace TheCenter.Project
 
             if (itemName == "mask")
             {
-                stopwatch.Stop();
-                int timeElapsed = stopwatch.Elapsed.Seconds;
+                countdown.stop();
+                //int timeElapsed = stopwatch.Elapsed.Seconds;
+                int timeElapsed = countdown.GetSeconds();
 
                 if (timeElapsed > 10)
                 {
@@ -181,12 +181,17 @@ namespace TheCenter.Project
             if (index > -1 && roomItem != null && OK == true)
             {
                 if (itemName == "booth")
-                {
+                { 
                     if (CurrentPlayer.Score >= 25)
                     {
                         CurrentPlayer.IncreaseScore(CurrentRoom.Items[index].Name.Length);
                         CurrentRoom.UseItem(roomItem);
                         Console.Clear();
+                        if (running)
+                        {
+                            myproc.Kill();
+                            running = false;
+                        }
                         Console.WriteLine("You've Won! This Level is COMPLETED.");
                         Console.WriteLine("Enjoy the ride down. The next level won't be so easy!");
                         Console.WriteLine("");
@@ -319,7 +324,18 @@ namespace TheCenter.Project
 
                 if (CurrentRoom.Name == "AirlockTreatment Room")
                 {
-                    stopwatch.Start();
+                    countdown.start();
+                    countdown.warningBeep();
+                }
+
+                if (CurrentRoom.Name == "Central Room")
+                {
+                    countdown.stop();
+                    if (!running)
+                    {
+                        myproc.Start();
+                        running = true;
+                    }
                 }
 
                 Console.WriteLine("");
@@ -453,7 +469,11 @@ namespace TheCenter.Project
 
         public void Quit()
         {
-            myproc.Kill();
+            if (running)
+            {
+                myproc.Kill();
+                running = false;
+            }
             Console.Clear();
             Environment.Exit(0);
         }
@@ -461,6 +481,18 @@ namespace TheCenter.Project
         public void YouDied()
         {
             Console.Clear();
+
+            if (running)
+            {
+                myproc.Kill();
+                running = false;
+            }
+
+            if (CurrentRoom.Name == "AirlockTreatment Room")
+            {
+                Console.WriteLine(". . . you realize it takes some time to properly put on a gas mask.\r\n\r\n");
+            }
+
             Console.WriteLine("The good news . . .\r\nyou don't have to worry about anything . . .EVER.\r\nThe bad news is . . . \r\n\r\nYOU DIED!\r\n\r\nPress Enter...");
             string res = Console.ReadLine();
             Console.Clear();
@@ -469,9 +501,13 @@ namespace TheCenter.Project
 
         public void Restart()
         {
-            myproc.Kill();
+            if (running)
+            {
+                myproc.Kill();
+                running = false;
+            }
+
             Console.Clear();
-            //Console.WriteLine("Restarting");
             CurrentPlayer.Inventory.Clear();
             CurrentPlayer.Score = 0;
             Setup();
